refactor(framework): move style module to core
